be/requirements.txt
flask
torch
ultralytics
numpy
opencv-python-headless
minio
psycopg2-binary
mlflow